7_lab/pyproject.toml
[project]
name = "7-lab"
version = "0.1.0"
description = "Створюємо першого АІ агента"
authors = [
    {name = "BobasB",email = "bugil.bogdan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "google-adk (>=1.27.2,<2.0.0)",
    "python-telegram-bot (>=22.7,<23.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
